feat: use Clerk Node SDK
